+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@ Please refer to the following release notes when upgrading your version of Knock
 
 ## 0.16.7
 
+* replaced `sortedIndex` by `comparator` for sorting a kb.collectionObservable
 * added extensibility to knockback-validation.js through data-bind on input elements. Added $enabled and $disabled and renamed internal helpers ($error_count, $valid, and $active_error) with '$' prefix.
 * published filters function on kb.CollectionObservable.
 * added stricted checking for compatible type setting on a kn.CollectionObservable's ko.observableArray.
@@ -10,6 +11,10 @@ Please refer to the following release notes when upgrading your version of Knock
 * made the kb.ModelWatcher into a general-purpose kb.EventWatcher so it could also watch collection events.
 * bug fix: added dispose of kb.CollectionObservable _mapping
 * bug fix: implemented general-purpose basic object test
+
+### Deprecations
+
+* **ko.collectionObservable sortedIndex has been replaced by comparator so you can resort in custom ways rather than reducing the sorting down to an index.
 
 ## 0.16.6